Limit number of procs to prevent high load notices
On systems with many cores/hyperthreading go spawns as many threads as it can. Due to the way load is calculated this can result in notices of high load as explained in https://blog.avast.com/investigation-of-regular-high-load-on-unused-machines-every-7-hours . To prevent these notices we limit the number of threads to max 4.